requirements.txt
streamlit==1.45.1
google-generativeai==0.8.5
sympy==1.14.0
python-dotenv==1.1.0